refactor: modular directory structure.
Migrated to a new directory structure that allows individual
functions/classes to be located inside their own private modules.

Removed `sphinx.ext.napoleon` because it was causing unpredictable
conflicts with `numpydoc`.

Updated autosummary templates to match scipy style.

BREAKING CHANGE: namespaces have moved around, which will break
existing import statements. `Item`, `Arrangement`, `Knapsack`, and
`Sampler` are now all imported from the `pykp.knapsack` namespace.
