+++ docs/source/_templates/autosummary/method.rst
@@ -1,0 +1,13 @@
+:orphan:
+
+.. raw:: html
+
+   <div class="prename">{{ module }}.{{ class }}.</div>
+   <div class="empty"></div>
+
+{{ name }}
+{{ underline }}
+
+.. currentmodule:: {{ module }}
+
+.. automethod:: {{ objname }}
