+++ docs/source/_templates/autosummary/property.rst
@@ -1,0 +1,13 @@
+:orphan:
+
+.. raw:: html
+
+   <div class="prename">{{ module }}.{{ class }}.</div>
+   <div class="empty"></div>
+
+{{ fullname }}
+{{ underline }}
+
+.. currentmodule:: {{ module }}
+
+.. autoproperty:: {{ objname }}
