--- docs/source/_templates/function.rst
@@ -1,5 +1,0 @@
-{{ fullname | escape | underline}}
-
-.. currentmodule:: {{ module }}
-
-.. autofunction:: {{ fullname }}
