+++ docs/source/conf.py
@@ -29,7 +29,7 @@ copyright = "2025, Hassan Andrabi"
 author = "Hassan Andrabi"
 
 extensions = [
-    "sphinx.ext.napoleon",
+    # "sphinx.ext.napoleon",
     "sphinx.ext.autodoc",
     "sphinx.ext.autosummary",
     "sphinx.ext.coverage",
@@ -56,24 +56,6 @@ numpydoc_use_plots = True
 
 # The name of the Pygments (syntax highlighting) style to use.
 pygments_style = "sphinx"
-
-# -----------------------------------------------------------------------------
-# Napoleon configuration
-# -----------------------------------------------------------------------------
-napoleon_google_docstring = False
-napoleon_numpy_docstring = True
-napoleon_include_init_with_doc = False
-napoleon_include_private_with_doc = False
-napoleon_include_special_with_doc = True
-napoleon_use_admonition_for_examples = False
-napoleon_use_admonition_for_notes = False
-napoleon_use_admonition_for_references = False
-napoleon_use_ivar = False
-napoleon_use_param = True
-napoleon_use_rtype = True
-napoleon_preprocess_types = False
-napoleon_type_aliases = None
-napoleon_attr_annotations = True
 
 # -----------------------------------------------------------------------------
 # Matplotlib plot_directive options
