+++ docs/source/reference/index.rst
@@ -5,12 +5,8 @@ This reference manual details functions, modules, and objects included in PyKP, 
 
 .. autosummary::
    :toctree: generated
-   :template: module.rst
    :recursive:
    
    pykp.knapsack
-   pykp.item
-   pykp.arrangement
-   pykp.sampler
    pykp.solvers
    pykp.metrics
