+++ pykp/__init__.py
@@ -1,20 +1,16 @@
-"""PyKP: A Python library for the Knapsack Problem."""
+"""
+PyKP: A Python library for the Knapsack Problem.
+================================================
 
-from .arrangement import Arrangement
-from .item import Item
-from .knapsack import Knapsack
-from .metrics import phase_transition, sahni_k
-from .sampler import Sampler
-from .solvers import branch_and_bound, greedy, minizinc
+Documentation is available in the docstrings and
+online at https://pykp.readthedocs.io/latest/index.html
 
-__all__ = [
-    "Arrangement",
-    "Item",
-    "Knapsack",
-    "Sampler",
-    "phase_transition",
-    "sahni_k",
-    "branch_and_bound",
-    "greedy",
-    "minizinc",
-]
+Subpackages
+-----------
+::
+
+ knapsack                     --- Core classes for the knapsack problem
+ metrics                      --- Metrics for evaluating instance complexity
+ solvers                      --- Solvers for the knapsack problem
+
+"""  # noqa: D205
