+++ pykp/metrics/__init__.py
@@ -1,0 +1,21 @@
+"""Metrics for evaluating knapsack instance complexity.
+
+.. currentmodule:: pykp.metrics
+
+PyKP ``metrics`` provides metrics to evaluate the complexity of knapsack
+problem instances.
+
+Metrics
+=======
+
+.. autosummary::
+   :toctree:
+
+   phase_transition - Phase transition of the knapsack problem.
+   sahni_k - Sahni-k metric for the knapsack problem.
+"""
+
+from ._phase_transition import phase_transition
+from ._sahni_k import sahni_k
+
+__all__ = ["phase_transition", "sahni_k"]
