+++ tests/test_knapsack.py
@@ -5,7 +5,7 @@ import pathlib
 import numpy as np
 import pytest
 
-from pykp import Item, Knapsack
+from pykp.knapsack import Item, Knapsack
 
 
 @pytest.fixture
